fix IPv4 data socket getting broken when CONFIG_IPV6=n
Similar to the management sockets introduced in the previous patch, we
have a problem with the receiver not being able to listen even for an
IPv4 address when IPv6 is compiled out of the kernel.

Fix the problem by extending the socket API with connection-less
functions, and also move the timestamping code from common.c there.

Fixes: 563b4d118668 ("rcv: properly support IPv6")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>