management: be more user-friendly with error messages
Currently we duplicate error messages for SET and for GET TLV actions,
which is confusing because we don't really know what went wrong.
Also, we print the TLV IDs as numbers, which is again confusing because
the user needs to look up the code.

Create individual error strings for GET and for SET, and strings for
each management TLV ID.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>