report: omit packets not completely TX timestamped from summary calculation
Like the dedicated sender, the isochron daemon may fail to timestamp
some packets. But as opposed to the dedicated sender, it will not
propagate this to the orchestrator, which will still retrieve the packet
logs and save them to a file.

When that happens, isochron report looks at a log file containing zeroes
for TX timestamps, and mistakenly shows a hugely incorrect path delay,
etc etc, when in fact it should just exclude those missing TX timestamps
and warn about what went wrong.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>