management: print TLV responses with empty payload as such
We count the management TLV response size including the TLV header
length itself, which results in confusing messages such as this:

Failed to query MID VID: expected payload length 12 in response, got 8 instead

where in fact, 8 bytes was the length of the TLV header, and the payload
size itself was 0. This corresponds to empty payloads sent by the other
end as result of an error.

Detect this and just print that the payload is empty.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>