daemon: forward test state based on whether TX timestamping thread is done
Currently we say that a test is done when the sender thread is done, but
the TX timestamping thread might still be doing work, and this is in a
race with the orchestrator.

Fix this by adding a test state in the sender itself, and tracking when
the TX timestamping thread exits.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>