sk: make TCP listen sockets VRF-aware
Introduce a --stats-address option in the programs with a management
socket (daemon, rcv) so that they can bind to a particular VRF
(--stats-address 0.0.0.0%vrf0). This is useful when these programs are
accessed over the network and the main kernel routing table is not aware
of a route to the destination IP address in the packets (kselftests).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>