sk: fix broken L2 TX timestamping
Currently, L2 TX timestamps are not received through the socket error
queue beyond a certain point (200 packets or so).

Seemingly, this is because we should bind(2) to the PF_PACKET socket if
we want to receive the error packets with the timestamping cmsgs.

We never called bind(2) on the L2 socket in the history of the sender
program, but we also passed "protocol" as IPPROTO_RAW to the socket(2)
call. I'm not quite clear how this is interpreted by net/packet/af_packet.c
or if that was even a valid thing to do in the first place, but fact is
that TX timestamping used to work in this unbound socket configuration
that way.

What the blamed commit below changed was it replaced IPPROTO_RAW for the
sender with htons(ethertype), in the attempt to unify the sender with
the receiver socket creation code (for the receiver we did pass the
EtherType as the PF_PACKET protocol).

It's not really the fault of the commit per se, it's more of a lingering
bug that got exposed just now.

This change reshapes the unconnected socket API again. It replaces
sk_l2() with sk_bind_l2() so that the sender actually binds to the
protocol and address, making it able to recvmsg() the timestamps.
Interestingly, UDP data sockets do *not* need to bind(2) to be able to
get the timestamps, in fact binding is not even possible due to the
kernel not being able to assign us the destination IP address and port.
So sk_udp() remains, but sk_bind_udp_any() is now reimplemented as
sk_bind_udp() which is a bound form of sk_udp(), so we have more common
code.

Since sk_bind_l2() needs the L2 address to bind to, and the caller
doesn't really need that in any way, the change also hides the struct
sk_addr inside sk.c and removes it from the API, making struct sk_msg be
attached directly to a socket as far as callers are concerned.

Fixes: fb5c44fad600 ("fix IPv4 data socket getting broken when CONFIG_IPV6=n")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>