orchestrate: don't collect logs after being interrupted
If we want to kill the orchestrator with SIGTERM or SIGINT, we want it
to actually stop, not continue saving partial logs to the filesystem.

There is a signal_received handler in prog_monitor_sync, but there we
just return false and not an error code. This makes us stop the
receivers due to the perceived loss of sync, but the program's unaltered
logic is to retry. So check for signal_received again in the outer loop
(prog_run_test) and return an actual error code to stop the program.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>