send: halve the time margins
Currently we have a hardcoded time margin of one second between the
sender starting its pthread and that pthread's first wakeup time.

In the orchestrator, we double this to two seconds, just to make sure
the sender will not need to auto-advance that wakeup time, and therefore
play nice with other coordinated senders.

But 2 seconds overhead for every test is a bit of a headache, and we'll
introduce a new check soon for the validity of the start time anyway.
So we are willing to be a bit more aggressive about reducing the margins
in half, considering that most of the time those will be enough, and
when they won't, we'll just retry the test.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>