syncmon: accelerate initial sync checking process
When orchestrating nodes as part of a kselftest and thousands of tests
are being run, typically it will take a while for synchronization to be
initially established, then it will be solid for good.

However, the pattern during kselftests will be to restart the senders
for each test to be run, and this means we need to go through 3 sync
checks, at 1 second apart each, before determining that we are good to
go.

In fact, the initial sync checking becomes the predominant factor that
incurs delay in such a kselftest. There is a desire to do something to
detect that we are already in sync and accelerate the initial waiting
process in that case.

There is a second problem caused by our hardcoded sleep(1) calls, which
is that during relatively short tests (shorted than 3 seconds), we
effectively do not have enough time to determine that the sync status
went bad, because of the hardcoded interval between checks.

So the situation calls for dynamic sync checking intervals
(1) at the beginning of the test
(2) during the test itself

This change initially started as just that - a determination of an
interval and initial_interval, plus what is now syncmon_init_ts() and
syncmon_next(). Those are the core of this change. The goal was to
abstract the sync monitoring intervals to some other place, which got
called "syncmon".

But for (1), it is required that we essentially go through a sync check
process just to determine how frequently to perform the sync checks.
This would result in a lot of code duplication, because we have
(different) sync monitoring code for the standalone sender and for the
orchestrator.

So what was initially just the code for sync interval calculation
gradually became the entire sync monitoring logic which now lies in
syncmon.c.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>