management: add error code and message propagation
Currently when a managed node fails to complete a GET or SET request, it
sends back a management with a RESPONSE action and an empty payload.

Add support in managed programs (receiver, daemon) to keep an error code
and error message (extack name is taken from netlink extended ack concept)
per MID, so that the management client can query these at a later time
if desired.

The error table isn't exposed to the managed programs themselves, but
kept inside management.c. All the managed programs have to do is to
return the desired error code in their (*get) and (*set) handlers, and
optionally set the extack. The generic management code automatically
takes care of forwarding these errors to the client, by responsing to a
newly introduced GET_ERROR TLV action.

Special attention that a GET and a SET handler for the same MID will
overwrite each other's error code and extack.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>