daemon: avoid freeing unallocated data socket when orchestrator disconnects
There is an unbalance created in prog_update_test_state() between
starting and stopping the session. isochron_send_init_data_sock() is
called when the session is started, but
isochron_send_teardown_data_sock() is called when the sender is torn
down, which is done also when the orchestrator disconnects.

The implication is that isochron_send_teardown_data_sock() can be called
even if isochron_send_init_data_sock() was never called in the first
place, and this leads to freeing data structures which were never
allocated.

To clarify what the code is trying to do:

- rename "test_running" to "session_active". A session remains active
  even after the test is no longer running, for the orchestrator to do
  things such as retrieving the logs.

- rename prog_prepare_session() to prog_prepare_send_session()

- create a counterpart of the above called prog_teardown_send_session()

Moving the session cleanup in prog_teardown_send_session() creates
another problem: currently the orchestrator tries to retrieve the logs
after it has called prog_stop_senders() which closed the session.

This would trigger invalid accesses to send->log in the daemon, so deny
that, and change the orchestrator code to call prog_collect_logs()
before calling prog_stop_senders().

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>