daemon: propagate errors from the sender threads to the orchestrator
Currently we have kselftests where there is so much data being collected
that it is infeasible to look at it all, and orchestrated tests that run
for so long that catching errors in person is similarly out of the
question.

With SO_TXTIME, senders are able to capture all sorts of errors, and
when no errors occur, the test can truly be considered to be reliable.

The problem appears when this is an orchestrated sender via a daemon,
because currently the sender is the one who keeps its internal test
state, and the daemon just forwards it to the orchestrator as either
TEST_STATE_IDLE or TEST_STATE_RUNNING. An idle state can also mean a
failed test, but the orchestrator doesn't know this, and collects the
logs as good.

Fix this by introducing a TEST_STATE_FAILED, and using the management
extack to propagate further information about what exactly failed.
Then modify the orchestrator to call GET_ERR on the TEST_STATE MID
before actually calling prog_stop_senders(), because that function calls
a SET action on the TEST_STATE MID, and that SET overwrites the extack
so we never get to see the error.

In the same idea, if the test is invalid for whatever reason, don't
bother to collect the logs from the orchestrated senders.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>