daemon: close race condition created by declaring the test done too soon
We say that a sender has done when the sender thread has ended, however
this opens the possibility for the orchestrator to collect the logs, and
the last few packets may have not received TX timestamps yet.

Declare that the test is done when the TX timestamp thread has finished
instead, which closes the race.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>