orchestrate: wait until sync is done when repeating a test
Currently we only wait for initial sync at the beginning of the first
test. Ironically, if there is a loss of sync during the test, we do not
wait again for the sync status to stabilize, but jump right ahead and
retry the test. That is likely to fail again.

Address this impulsive behavior by moving the syncmon_wait_until_ok()
function call inside the do {} while() loop.

At the same time, move the syncmon_init() call within the loop, because
we want to update the sync check intervals after a test failure -
probably to slower ones.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>