ptpmon: make it clear that the UDS paths are length-bounded
As explained in commit dfd5c9156da0 ("isochron: avoid use of strncpy"),
ptpmon_create() already length checks the uds_local and uds_remote
strings before saving them and using them in uds_send() and uds_recv().

Yet, static analysis tools do not see this, and complain that the
uds_remote argument is of an unknown length (which it is not).
Stick it in their face by specifying the length of the array as the
actual argument of the function.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>