management: check for truncation when marshalling strings
Validate the length of interface name and UNIX domain socket path
strings before saving them to program data. Do this by using a bounce
buffer that is one byte larger than the final destination, and check the
string length within the bounce buffer first. Otherwise, we may end up
accessing data beyond the source, because the source and destination
strings are of the same length.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>