sk: change return code of sk_send and sk_recv
Currently we have the "if (len <= 0) return len ? len : -ECONNRESET"
pattern spread out in many places that call sk_recv and sk_send, but not
everywhere.

Furthermore, we may miss a socket disconnection event, because we only
check for that in relatively few places. The bug that this causes is
that if the management socket dies in a few key places,
isochron_mgmt_event() will return -ECONNRESET rather than the orderly 0
plus socket_closed = 0, as it should. In turn, this makes the receiver
and/or daemon think there is an error, and exit prog_mgmt_loop() instead
of continuing to listen for other connections.

It's clear that there needs to be a more robust mechanism to interact
with the socket API, and that is to make sk_send and sk_recv return just
0 on success, and a negative return code on any failure (including
socket disconnection). When the socket disconnects, sock->closed will be
set to true.

We want to suppress error printing when the cause of the failure is a
socket disconnect event, so we introduce a new sk_err() macro which
checks whether the socket was disconnected first.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>