completion: add support for symlinked functions
Update the bash-completion script such that it properly sets the "prog"
and "subprog" variables even in the presence of symlinked isochron
binaries. Since bash will look for a completion file equal to the name
of the binary, this also means we need to symlink the completion script
for each isochron function.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>