send: avoid NULL pointer dereference when --client is omitted
Currently we have a SIGSEGV when running the following command, due to a
NULL "sk" pointer in sk_send():

isochron send --interface eno0 --cycle-time 0.001 --frame-size 900 \
	--omit-sync --num-frames 100 --priority 7 --vid 0 --no-ts \
	--dmac 01:80:c2:00:00:0e

Access not within mapped region at address 0x4
  at 0x121890: sk_send (sk.c:368)
  by 0x11284F: isochron_send_tlv (management.c:136)
  by 0x1130A7: isochron_query_mid (management.c:348)
  by 0x12365B: syncmon_node_query_link_state_remote (syncmon.c:72)
  by 0x123777: syncmon_node_update_link_state (syncmon.c:110)
  by 0x1240EB: syncmon_sync_ok (syncmon.c:363)
  by 0x124387: syncmon_init_initial_interval (syncmon.c:426)
  by 0x1244D7: syncmon_init (syncmon.c:459)
  by 0x11E2DB: prog_init_syncmon (send.c:576)
  by 0x11FA13: prog_init (send.c:1248)
  by 0x120CBF: isochron_send_main (send.c:1834)
  by 0x10FB83: main (isochron.c:167)

This is because prog->mgmt_sock is NULL, since that only gets created
when we have a connection specified using --client. Nonetheless, we
create a remote syncmon node using a socket we don't have, and that
ultimately queries a receiver we can't reach.

Fixes: 3b7df52ee844 ("syncmon: accelerate initial sync checking process")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>