man: isochron-send: clarify --frame-size relationship to IEEE
Linux packets don't contain an FCS, this is always computed by the NIC,
so isochron's understanding of "frame size" is more or less "buffer
size" when using an L2 transport (PF_PACKET socket), although this frame
size is consistent when using AF_INET/AF_INET6 sockets.

Add a note stating that the IEEE definition of this term is different
than isochron's understanding of it.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>