rcv: allow sysmon to be skipped on interfaces without a PHC
It might make sense to use isochron on i.e. the veth driver.
But sysmon_create() -> phc_open() fails (because there is no PHC), and
the receiver process fails to start.

It makes sense to ignore the failure if --omit-hwts is used, and simply
refuse at runtime any ISOCHRON_MID_SYSMON_OFFSET management requests
instead. It is likely that the sender will also use --omit-hwts if we
are talking about veth pairs, but it can also use --omit-remote-sync.
So things might work out.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>